refactor with useCanEdit hook
