+++ components/item-info.js
@@ -6,7 +6,7 @@ import Dropdown from 'react-bootstrap/Dropdown'
 import Countdown from './countdown'
 import { abbrNum, numWithUnits } from '@/lib/format'
 import { newComments, commentsViewedAt } from '@/lib/new-comments'
-import { datePivot, timeSince } from '@/lib/time'
+import { timeSince } from '@/lib/time'
 import { DeleteDropdownItem } from './delete'
 import styles from './item.module.css'
 import { useMe } from './me'
@@ -28,6 +28,7 @@ import { useToast } from './toast'
 import { useShowModal } from './modal'
 import classNames from 'classnames'
 import SubPopover from './sub-popover'
+import useCanEdit from './use-can-edit'
 
 export default function ItemInfo ({
   item, full, commentsText = 'comments',
@@ -35,33 +36,18 @@ export default function ItemInfo ({
   onQuoteReply, extraBadges, nested, pinnable, showActionDropdown = true, showUser = true,
   setDisableRetry, disableRetry
 }) {
-  const editThreshold = datePivot(new Date(item.invoice?.confirmedAt ?? item.createdAt), { minutes: 10 })
   const { me } = useMe()
   const router = useRouter()
   const [hasNewComments, setHasNewComments] = useState(false)
   const root = useRoot()
   const sub = item?.sub || root?.sub
+  const [canEdit, setCanEdit, editThreshold] = useCanEdit(item)
 
   useEffect(() => {
     if (!full) {
       setHasNewComments(newComments(item))
     }
   }, [item])
-
-  // deleted items can never be edited and every item has a 10 minute edit window
-  // except bios, they can always be edited but they should never show the countdown
-  const noEdit = !!item.deletedAt || (Date.now() >= editThreshold) || item.bio
-  const authorEdit = me && item.mine
-  const [canEdit, setCanEdit] = useState(!noEdit && authorEdit)
-
-  useEffect(() => {
-    // allow anon edits if they have the correct hmac for the item invoice
-    // (the server will verify the hmac)
-    const invParams = window.localStorage.getItem(`item:${item.id}:hash:hmac`)
-    const anonEdit = !!invParams && !me && Number(item.user.id) === USER_ID.anon
-    // anonEdit should not override canEdit, but only allow edits if they aren't already allowed
-    setCanEdit(canEdit => canEdit || anonEdit)
-  }, [])
 
   // territory founders can pin any post in their territory
   // and OPs can pin any root reply in their post
